+++ kibot/out_base_3d.py
@@ -190,7 +190,7 @@ class Base3DOptions(VariantOptions):
                     models.add(full_name)
         return list(models)
 
-    def filter_components(self):
+    def filter_components(self, highlight=None):
         if not self._comps:
             # No variant/filter to apply
             if self.download_models():
@@ -201,7 +201,7 @@ class Base3DOptions(VariantOptions):
                 self.undo_3d_models_rename(GS.board)
                 return ret
             return GS.pcb_file
-        self.filter_pcb_components(GS.board, do_3D=True, do_2D=True)
+        self.filter_pcb_components(GS.board, do_3D=True, do_2D=True, highlight=highlight)
         self.download_models()
         fname = self.save_tmp_board()
         self.unfilter_pcb_components(GS.board, do_3D=True, do_2D=True)